% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{angle_in_sector}
\alias{angle_in_sector}
\title{Determine if an angle lies within a sector}
\usage{
angle_in_sector(x, start, end, degrees = TRUE)
}
\arguments{
\item{x}{One or more angles to check.}

\item{start}{The start angle of the sector}

\item{end}{The end angle of the sector}

\item{degrees}{(logical) If \code{TRUE} (default), all values are treated as
angles in degrees. Set to \code{FALSE} if values should be treated as
radians.}
}
\value{
\code{TRUE} if the angle lies within the sector.
}
\description{
The sector is defined as the angular range proceeding clockwise from the
start angle to the end angle.
}
\details{
It is up to the user to ensure that the test and sector angles are either all
compass bearings or all Cartesian angles. Terrible things will happen
otherwise.
}
