% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{compass2cartesian}
\alias{compass2cartesian}
\title{Convert a compass bearing to a Cartesian angle}
\usage{
compass2cartesian(x, degrees = TRUE)
}
\arguments{
\item{x}{A numeric vector of one or more compass bearings. By default, it is
assumed that the angular units are degrees. See the \code{degrees} argument
if input values should be treated as radians.}

\item{degrees}{(logical) If \code{TRUE} (default), input values are treated
as compass angles in degrees. Set to \code{FALSE} if input values should be
treated as radians.}
}
\value{
A numeric vector of corresponding Cartesian angles, with angular
units the same as those of the input values. The returned values will be in
the range \code{[0, 360)} when units are degrees, or \code{[0, 2*pi)} when
units are radians.
}
\description{
Compass bearings range from north (0 degrees) through east (90), south (180)
and west (270). In contrast, Cartesian angles define 0 degrees as the
direction of the positive X-axis and then proceed counter-clockwise such that
90 degrees (pi/2 radians) is the direction of the positive Y-axis etc.
}
\examples{
# Convert a north-west compass bearing (315 degrees) to the corresponding
# Cartesian angle expressed in radians (3*pi/4 = 2.356194)
#
compass2cartesian(315) |> deg2rad()

}
\seealso{
\code{\link[=cartesian2compass]{cartesian2compass()}}
}
