% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert angular values from degrees to radians}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{A numeric vector of one or more finite angular values expressed in
degrees.}
}
\value{
A numeric vector of angular values in radians.
}
\description{
All input values must be finite. Missing values are allowed and will simply
be returned as \code{NA}.
}
