% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{get_compass_direction}
\alias{get_compass_direction}
\title{Determine the compass bearing from one point to another}
\usage{
get_compass_direction(p0, p1)
}
\arguments{
\item{p0}{A two-element numeric vector of X-Y coordinates for the reference
point.}

\item{p1}{A two-element numeric vector of X-Y coordinates for the query
point.}
}
\value{
The compass bearing of p1 from p0 expressed in degrees, or \code{NA}
if the two points are the same.
}
\description{
Given a reference point p0 and a query point p1, each represented by a
two-element vector of X:Y coordinates, find the compass bearing in degrees of
p1 from reference point p0.
}
\examples{
p0 <- c(700000, 6530000)
p1 <- c(695000, 6535000) # a point north-west of p0
get_compass_direction(p0, p1)  # returns 315 degrees

# If the two points are the same there is no valid direction
get_compass_direction(p0, p0)  # returns NA

}
