% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{get_shortest_cartesian_rotation}
\alias{get_shortest_cartesian_rotation}
\title{Find the shortest angle of rotation between Cartesian angles}
\usage{
get_shortest_cartesian_rotation(a0, ax)
}
\arguments{
\item{a0}{The reference compass bearing (radians).}

\item{ax}{One or more bearings to compare to \code{a0} (radians).}
}
\value{
A numeric vector the same length as \code{ax} with values for the
angle of shortest rotation expressed in radians.
}
\description{
Given a reference angle \code{a0} and one or more angles
\code{ax}, this function determines the shortest angle of rotation from
\code{a0} to each \code{ax}. Positive return values indicate clockwise
rotations, while negative values indicate counter-clockwise rotations. A
positive value will be returned for the \code{pi} radians rotation between
two opposite bearings.
}
\details{
Adapted from a post by Allan Cameron on
StackOverflow: https://stackoverflow.com/a/63429855/40246.
}
