% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{get_shortest_compass_rotation}
\alias{get_shortest_compass_rotation}
\title{Find the shortest angle of rotation between compass bearings}
\usage{
get_shortest_compass_rotation(b0, bx)
}
\arguments{
\item{b0}{The reference compass bearing (degrees).}

\item{bx}{One or more bearings to compare to \code{b0} (degrees).}
}
\value{
A numeric vector the same length as \code{bx} with values for the
angle of shortest rotation expressed in degrees.
}
\description{
Given a reference compass bearing \code{b0} and one or more bearings
\code{bx}, this function determines the shortest angle of rotation from
\code{b0} to each \code{bx}. Positive return values indicate clockwise
rotations, while negative values indicate counter-clockwise rotations. A
positive value will be returned for the 180 degree rotation between two
opposite bearings.
}
\details{
Adapted from a post by Allan Cameron on
StackOverflow: https://stackoverflow.com/a/63429855/40246.
}
\examples{
# north-west to north-east
get_shortest_compass_rotation(315, 45)  # return 90 (clockwise rotation)

# north-east to north-west
get_shortest_compass_rotation(45, 315)  # returns -90 (anti-clockwise rotation)

# Exactly opposite bearings are treated as a clockwise rotation (positive value)
get_shortest_compass_rotation(270, 90)  # returns 180

# Input bearings outside the usual compass range (0 <= b < 360 degrees) are
# permitted. For example, the query below will return -135 for the counter-
# clockwise rotation from north-west (coded as -45 degrees) to south
#
get_shortest_compass_rotation(-45, 180) # returns -135

}
