% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{get_target_point}
\alias{get_target_point}
\title{Find a target point at a given bearing and distance from a reference point}
\usage{
get_target_point(p0, bearing, distance, degrees = TRUE)
}
\arguments{
\item{p0}{The reference point. Either a two-element vector of X-Y
coordinates, an \code{sf} point geometry object, a single element from an
\code{sfc} point geometry list or a single record from an \code{sf} spatial
data frame.}

\item{bearing}{The compass direction from p0 to the target point.}

\item{distance}{Distance to the new point. Generally, this will be a positive
value. If zero, the function simply returns the reference point
coordinates. If negative, a point in the opposite direction to that
specified by the \code{bearing} will be returned.}

\item{degrees}{(logical) If \code{TRUE} (default), all bearings are treated
as compass angles in degrees. Set to \code{FALSE} if values should be
treated as radians.}
}
\value{
A two-element vector with the coordinates of the target point.
}
\description{
Given a reference point \code{p0}, find the coordinates of a target point
located at a given compass bearing and distance from \code{p0}.
}
\note{
This function currently only useful for points in projected coordinate
systems, not geographic (longitude, latitude) coordinates.
}
\examples{
# Reference point coordinates (assuming map units are metres)
p0 <- c(305170, 6190800)

# Find the target point 170 metres from the reference point along a compass
# bearing of 10 degrees (slightly east of north)
#
ptarget <- get_target_point(p0, bearing = 10, distance = 170)
cat("Target point coordinates:", ptarget)

}
