% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_sf_object}
\alias{is_sf_object}
\title{Test if an arbitrary object is some type of simple feature}
\usage{
is_sf_object(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
\code{TRUE} if \code{x} is a simple feature object of some kind.
}
\description{
Tests if an object is one of the simple feature types known to the \code{sf}
package. It does this by attempting to query the geometry type of the object
via \code{\link[sf]{st_geometry_type}} and checking whether this results in
an error (indicating a non-sf object).
}
\examples{
coords <- c(305170,	6190800)
p <- sf::st_point(coords)

is_sf_object(coords)  # FALSE
is_sf_object(p)  #TRUE

}
