% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{is_upwind}
\alias{is_upwind}
\title{Determine if point locations are up-wind of a reference point}
\usage{
is_upwind(p0, pquery, wdir_degrees, halfspan = 90, strict = FALSE)
}
\arguments{
\item{p0}{The reference point. Either a two-element vector of X-Y
coordinates, an \code{sf} point geometry object, a single element from an
\code{sfc} point geometry list or a single record from an \code{sf} spatial
data frame.}

\item{pquery}{One or more query points. If there is only query point it can
be provided in the same forms as described for \code{p0}. Multiple query
points can be provided as a matrix of X-Y coordinates, an \code{sfc} point
geometry list or an \code{sf} spatial data frame.}

\item{wdir_degrees}{Wind direction (single numeric value), expressed using
the standard meteorological convention as the compass direction from which
the wind is coming, e.g. 270 degrees means a westerly wind with air flow
from west to east.}

\item{halfspan}{The angular range (degrees) either side of the wind angle for
a bearing to be considered up-wind. Must be in the range \code{[0, 180]}.
The default value is 90 degrees for a semi-circle centred on the wind
direction.}

\item{strict}{(logical; default FALSE) Whether bearings must be strictly
within the angular range defined by \code{halfspan}. See
\code{\link{angle_in_range}} for more explanation.}
}
\description{
Given a single reference point location \code{p0} and one or more query point
locations \code{pquery}, together with a prevailing wind direction, determine
whether each query point is up-wind of the reference point. This function is
simply a short-cut to calling functions \code{get_compass_bearing()} and
\code{angle_in_range()}. To be up-wind, a query point must lie within an
angular range centred on the direction from which the wind is coming. By
default, the range is set to 90 degrees either side of the wind direction,
but this can be decreased for a stricter interpretation of 'up-wind'. To
accord with the standard meteorological convention, the wind direction value
\code{wdir_degrees} should always be expressed as a compass bearing in
degrees, representing the direction from which the wind is blowing, e.g. 270
degrees for a westerly wind with air flow from west to east.
}
\details{
Note that this function \strong{assumes} that wind direction, \code{wdir},
and the \code{halfspan} arguments are both expressed in degrees.
}
