% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{point_is_upwind}
\alias{point_is_upwind}
\title{Determine if a point location is up-wind of a reference point}
\usage{
point_is_upwind(p0, pquery, wdir_degrees, halfspan = 90)
}
\arguments{
\item{p0}{The reference point. Either a two-element vector of X-Y
coordinates, an \code{sf} point geometry object, a single element from an
\code{sfc} point geometry list or a single record from an \code{sf} spatial
data frame.}

\item{pquery}{One or more query points. If there is only query point it can
be provided in the same forms as described for \code{p0}. Multiple query
points can be provided as a matrix of X-Y coordinates, an \code{sfc} point
geometry list or an \code{sf} spatial data frame.}

\item{wdir_degrees}{Wind direction (single numeric value), expressed using
the standard meteorological convention as the compass direction from which
the wind is coming, e.g. 270 degrees means a westerly wind with air flow
from west to east.}

\item{halfspan}{The angular range (degrees) either side of the wind angle for
a bearing to be considered up-wind. Must be in the range \code{[0, 180]}.
The default value is 90 degrees for a semi-circle centred on the wind
direction.}
}
\description{
Given a reference point location \code{p0} and one or more query point
locations \code{pquery}, together with a prevailing wind direction, this
function determines whether each query point is up-wind of the reference
point. To be up-wind, a query point must lie strictly within an angular range
centred on the direction from which the wind is coming. By default, the range
is set to 90 degrees either side of the wind direction, but this can be
decreased for a stricter interpretation of 'up-wind'. To accord with the
standard meteorological convention, the wind direction value
\code{wdir_degrees} should always be expressed as a compass bearing in
degrees, representing the direction from which the wind is blowing, e.g. 270
degrees for a westerly wind with air flow from west to east.
}
\details{
Note that this function \strong{assumes} that wind direction and the range
up-wind bearings are both expressed in degrees.
}
