% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert angular values from radians to degrees}
\usage{
rad2deg(x)
}
\arguments{
\item{x}{A numeric vector of one or more finite angular values expressed in
radians.}
}
\value{
A numeric vector of angular values in degrees.
}
\description{
All input values must be finite. Missing values are allowed and will simply
be returned as \code{NA}.
}
