% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_functions.R
\name{is_upwind}
\alias{is_upwind}
\title{Determine if a location is up-wind of a reference point}
\usage{
is_upwind(p0, pquery, wdir_degrees, sector_width = 180, compass = TRUE)
}
\arguments{
\item{p0}{The reference point. Either a two-element vector of X-Y
coordinates, an \code{sf} point geometry object, a single element from an
\code{sfc} point geometry list or a single record from an \code{sf} spatial
data frame.}

\item{pquery}{One or more query points. If there is only query point it can
be provided in the same forms as described for \code{p0}. Multiple query
points can be provided as a matrix of X-Y coordinates, an \code{sfc} point
geometry list or an \code{sf} spatial data frame.}

\item{wdir_degrees}{Wind direction (single numeric value), expressed using
the standard meteorological convention as the compass direction from which
the wind is coming, e.g. 270 degrees means a westerly wind with air flow
from west to east.}

\item{sector_width}{Angular width (degrees) of the query sector. The default
and maximum allowable value is 180.}
}
\description{
Given a reference point location \code{p0} and one or more query point
locations \code{pquery}, together with a prevailing wind direction, this
function determines whether each query point is up-wind of the reference
point. To be up-wind, a query point must lie strictly within an angular
sector centred on the direction from which the wind is coming. By default,
the sector width is 180 degrees (i.e. 90 degrees either side of the wind
direction) but this can be decreased if a stricter interpretation of
'up-wind' is desired.
}
\details{
To accord with the standard meteorological convention, the wind direction
argument \code{wdir} should always be expressed as a compass bearing in
degrees, representing the direction from which the wind is blowing, e.g. 270
degrees for a westerly wind with air flow from west to east.
}
